requirements.txt
commentjson
imageio
numpy
numpy-quaternion
pybind11
pyquaternion
scipy
tqdm
mediapipe

